Introduce configs for optimizers and lr schedulers
