+++ src/fairseq2/generation/__init__.py
@@ -6,18 +6,27 @@
 
 from __future__ import annotations
 
-from fairseq2.generation.beam_search import (
-    BeamSearchSeq2SeqGenerator as BeamSearchSeq2SeqGenerator,
-)
-from fairseq2.generation.beam_search import (
-    BeamSearchSequenceGenerator as BeamSearchSequenceGenerator,
-)
-from fairseq2.generation.beam_search_algorithm import (
+from fairseq2.generation.beam_search.algo import (
     BeamSearchAlgorithm as BeamSearchAlgorithm,
 )
-from fairseq2.generation.beam_search_algorithm import BeamStep as BeamStep
-from fairseq2.generation.beam_search_algorithm import (
+from fairseq2.generation.beam_search.algo import BeamStep as BeamStep
+from fairseq2.generation.beam_search.algo import (
     StandardBeamSearchAlgorithm as StandardBeamSearchAlgorithm,
+)
+from fairseq2.generation.beam_search.factory import (
+    StandardBeamSearchConfig as StandardBeamSearchConfig,
+)
+from fairseq2.generation.beam_search.factory import (
+    beam_search_factories as beam_search_factories,
+)
+from fairseq2.generation.beam_search.factory import (
+    beam_search_factory as beam_search_factory,
+)
+from fairseq2.generation.beam_search.generator import (
+    BeamSearchSeq2SeqGenerator as BeamSearchSeq2SeqGenerator,
+)
+from fairseq2.generation.beam_search.generator import (
+    BeamSearchSequenceGenerator as BeamSearchSequenceGenerator,
 )
 from fairseq2.generation.chatbot import AbstractChatbot as AbstractChatbot
 from fairseq2.generation.chatbot import Chatbot as Chatbot
@@ -25,7 +34,31 @@ from fairseq2.generation.chatbot import ChatbotFactory as ChatbotFactory
 from fairseq2.generation.chatbot import ChatDialog as ChatDialog
 from fairseq2.generation.chatbot import ChatMessage as ChatMessage
 from fairseq2.generation.chatbot import chatbot_factories as chatbot_factories
-from fairseq2.generation.chatbot import register_chatbot as register_chatbot
+from fairseq2.generation.chatbot import chatbot_factory as chatbot_factory
+from fairseq2.generation.factory import BeamSearchConfig as BeamSearchConfig
+from fairseq2.generation.factory import SamplingConfig as SamplingConfig
+from fairseq2.generation.factory import (
+    create_beam_search_seq2seq_generator as create_beam_search_seq2seq_generator,
+)
+from fairseq2.generation.factory import (
+    create_beam_search_seq_generator as create_beam_search_seq_generator,
+)
+from fairseq2.generation.factory import (
+    create_sampling_seq2seq_generator as create_sampling_seq2seq_generator,
+)
+from fairseq2.generation.factory import (
+    create_sampling_seq_generator as create_sampling_seq_generator,
+)
+from fairseq2.generation.factory import (
+    seq2seq_generator_factories as seq2seq_generator_factories,
+)
+from fairseq2.generation.factory import (
+    seq2seq_generator_factory as seq2seq_generator_factory,
+)
+from fairseq2.generation.factory import (
+    seq_generator_factories as seq_generator_factories,
+)
+from fairseq2.generation.factory import seq_generator_factory as seq_generator_factory
 from fairseq2.generation.generator import (
     AbstractSeq2SeqGenerator as AbstractSeq2SeqGenerator,
 )
@@ -42,15 +75,25 @@ from fairseq2.generation.generator import (
     SequenceGeneratorOutput as SequenceGeneratorOutput,
 )
 from fairseq2.generation.generator import StepHook as StepHook
-from fairseq2.generation.sampler import Sampler as Sampler
-from fairseq2.generation.sampler import TopKSampler as TopKSampler
-from fairseq2.generation.sampler import TopPSampler as TopPSampler
-from fairseq2.generation.sampling import (
+from fairseq2.generation.sampling.factory import TopKSamplerConfig as TopKSamplerConfig
+from fairseq2.generation.sampling.factory import TopPSamplerConfig as TopPSamplerConfig
+from fairseq2.generation.sampling.factory import (
+    create_top_k_sampler as create_top_k_sampler,
+)
+from fairseq2.generation.sampling.factory import (
+    create_top_p_sampler as create_top_p_sampler,
+)
+from fairseq2.generation.sampling.factory import sampler_factories as sampler_factories
+from fairseq2.generation.sampling.factory import sampler_factory as sampler_factory
+from fairseq2.generation.sampling.generator import (
     SamplingSeq2SeqGenerator as SamplingSeq2SeqGenerator,
 )
-from fairseq2.generation.sampling import (
+from fairseq2.generation.sampling.generator import (
     SamplingSequenceGenerator as SamplingSequenceGenerator,
 )
+from fairseq2.generation.sampling.sampler import Sampler as Sampler
+from fairseq2.generation.sampling.sampler import TopKSampler as TopKSampler
+from fairseq2.generation.sampling.sampler import TopPSampler as TopPSampler
 from fairseq2.generation.step_processor import (
     BannedSequenceProcessor as BannedSequenceProcessor,
 )
