+++ src/fairseq2/generation/factory.py
@@ -7,26 +7,41 @@
 from __future__ import annotations
 
 from dataclasses import dataclass, field
-from typing import TYPE_CHECKING, Optional
+from typing import TYPE_CHECKING, Optional, Union
 
 from fairseq2.factory_registry import ConfigBoundFactoryRegistry
-from fairseq2.generation.beam_search import BeamSearchSequenceGenerator
-from fairseq2.generation.beam_search_algorithm import beam_search_factories
-from fairseq2.generation.generator import SequenceGenerator
-from fairseq2.generation.sampler import TopPSamplerConfig, sampler_factories
-from fairseq2.generation.sampling import SamplingSequenceGenerator
+from fairseq2.generation.beam_search.factory import beam_search_factories
+from fairseq2.generation.beam_search.generator import (
+    BeamSearchSeq2SeqGenerator,
+    BeamSearchSequenceGenerator,
+)
+from fairseq2.generation.generator import Seq2SeqGenerator, SequenceGenerator
+from fairseq2.generation.sampling.factory import TopPSamplerConfig, sampler_factories
+from fairseq2.generation.sampling.generator import (
+    SamplingSeq2SeqGenerator,
+    SamplingSequenceGenerator,
+)
 from fairseq2.models.decoder import DecoderModel
+from fairseq2.models.encoder_decoder import EncoderDecoderModel
 from fairseq2.typing import DataClass
 
 if TYPE_CHECKING:  # compat: remove when Python 3.9 support is dropped.
-    generator_factories = ConfigBoundFactoryRegistry[
+    seq_generator_factories = ConfigBoundFactoryRegistry[
         [DecoderModel], SequenceGenerator
     ]()
 else:
-    generator_factories = ConfigBoundFactoryRegistry()
+    seq_generator_factories = ConfigBoundFactoryRegistry()
 
+seq_generator_factory = seq_generator_factories.decorator
 
-register_generator = generator_factories.decorator
+if TYPE_CHECKING:  # compat: remove when Python 3.9 support is dropped.
+    seq2seq_generator_factories = ConfigBoundFactoryRegistry[
+        [EncoderDecoderModel], Seq2SeqGenerator
+    ]()
+else:
+    seq2seq_generator_factories = ConfigBoundFactoryRegistry()
+
+seq2seq_generator_factory = seq2seq_generator_factories.decorator
 
 
 @dataclass
@@ -44,7 +59,7 @@ class SamplingConfig:
     min_gen_len: int = 1
     """The minimum generation length."""
 
-    max_gen_len: int = 2048
+    max_gen_len: Union[int, tuple[int, int]] = 2048
     """The maximum generation length."""
 
     max_seq_len: Optional[int] = None
@@ -75,8 +90,8 @@ class SamplingConfig:
     """The sequence length capacity will be incremented by multiplies of this value."""
 
 
-@register_generator("sampling")
-def create_sampling_generator(
+@seq_generator_factory("sampling")
+def create_sampling_seq_generator(
     config: SamplingConfig, model: DecoderModel
 ) -> SamplingSequenceGenerator:
     try:
@@ -88,11 +103,54 @@ def create_sampling_generator(
             "The sampler cannot be created. See nested exception for details."
         ) from ex
 
+    if isinstance(config.max_gen_len, int):
+        max_gen_len = config.max_gen_len
+    else:
+        if config.max_gen_len[0] != 1:
+            raise ValueError("`max_gen_len` must be an integer.")
+
+        max_gen_len = config.max_gen_len[1]
+
     return SamplingSequenceGenerator(
         model,
         sampler,
         min_gen_len=config.min_gen_len,
-        max_gen_len=config.max_gen_len,
+        max_gen_len=max_gen_len,
+        max_seq_len=config.max_seq_len,
+        echo_prompt=config.echo_prompt,
+        compute_scores=config.compute_scores,
+        normalize_scores=config.normalize_scores,
+        temperature=config.temperature,
+        unk_penalty=config.unk_penalty,
+        len_penalty=config.len_penalty,
+        prefill_chunk_size=config.prefill_chunk_size,
+        decode_capacity_increment=config.decode_capacity_increment,
+    )
+
+
+@seq2seq_generator_factory("sampling")
+def create_sampling_seq2seq_generator(
+    config: SamplingConfig, model: EncoderDecoderModel
+) -> SamplingSeq2SeqGenerator:
+    try:
+        sampler_factory = sampler_factories.get(config.sampler, config.sampler_config)
+
+        sampler = sampler_factory()
+    except ValueError as ex:
+        raise ValueError(
+            "The sampler cannot be created. See nested exception for details."
+        ) from ex
+
+    max_gen_len = config.max_gen_len
+
+    if isinstance(max_gen_len, int):
+        max_gen_len = (1, max_gen_len)
+
+    return SamplingSeq2SeqGenerator(
+        model,
+        sampler,
+        min_gen_len=config.min_gen_len,
+        max_gen_len=max_gen_len,
         max_seq_len=config.max_seq_len,
         echo_prompt=config.echo_prompt,
         compute_scores=config.compute_scores,
@@ -121,7 +179,7 @@ class BeamSearchConfig:
     min_gen_len: int = 1
     """The minimum generation length."""
 
-    max_gen_len: int = 2048
+    max_gen_len: Union[int, tuple[int, int]] = 2048
     """The maximum generation length."""
 
     max_seq_len: Optional[int] = None
@@ -149,27 +207,73 @@ class BeamSearchConfig:
     """The sequence length capacity will be incremented by multiplies of this value."""
 
 
-@register_generator("beam_search")
-def create_beam_search_generator(
+@seq_generator_factory("beam_search")
+def create_beam_search_seq_generator(
     config: BeamSearchConfig, model: DecoderModel
 ) -> BeamSearchSequenceGenerator:
     try:
-        algorithm_factory = beam_search_factories.get(
+        beam_search_factory = beam_search_factories.get(
             config.algorithm, config.algorithm_config
         )
 
-        algorithm = algorithm_factory()
+        beam_search = beam_search_factory()
     except ValueError as ex:
         raise ValueError(
             "The beam-search algorithm cannot be created. See nested exception for details."
         ) from ex
 
+    if isinstance(config.max_gen_len, int):
+        max_gen_len = config.max_gen_len
+    else:
+        if config.max_gen_len[0] != 1:
+            raise ValueError("`max_gen_len` must be an integer.")
+
+        max_gen_len = config.max_gen_len[1]
+
     return BeamSearchSequenceGenerator(
         model,
-        algorithm=algorithm,
+        algorithm=beam_search,
         beam_size=config.beam_size,
         min_gen_len=config.min_gen_len,
-        max_gen_len=config.max_gen_len,
+        max_gen_len=max_gen_len,
+        max_seq_len=config.max_seq_len,
+        echo_prompt=config.echo_prompt,
+        normalize_scores=config.normalize_scores,
+        temperature=config.temperature,
+        unk_penalty=config.unk_penalty,
+        len_penalty=config.len_penalty,
+        prefill_chunk_size=config.prefill_chunk_size,
+        decode_capacity_increment=config.decode_capacity_increment,
+    )
+
+
+@seq2seq_generator_factory("beam_search")
+def create_beam_search_seq2seq_generator(
+    config: BeamSearchConfig, model: EncoderDecoderModel
+) -> BeamSearchSeq2SeqGenerator:
+    try:
+        beam_search_factory = beam_search_factories.get(
+            config.algorithm, config.algorithm_config
+        )
+
+        beam_search = beam_search_factory()
+    except ValueError as ex:
+        raise ValueError(
+            "The beam-search algorithm cannot be created. See nested exception for details."
+        ) from ex
+
+    max_gen_len = config.max_gen_len
+
+    if isinstance(max_gen_len, int):
+        max_gen_len = (1, max_gen_len)
+
+    return BeamSearchSeq2SeqGenerator(
+        model,
+        algorithm=beam_search,
+        beam_size=config.beam_size,
+        min_gen_len=config.min_gen_len,
+        max_gen_len=max_gen_len,
+        max_seq_len=config.max_seq_len,
         echo_prompt=config.echo_prompt,
         normalize_scores=config.normalize_scores,
         temperature=config.temperature,
