+++ src/fairseq2/models/__init__.py
@@ -12,6 +12,7 @@ from fairseq2.models.config_loader import (
 )
 from fairseq2.models.config_loader import get_model_family as get_model_family
 from fairseq2.models.config_loader import is_model_card as is_model_card
+from fairseq2.models.factory import model_factories as model_factories
 from fairseq2.models.loader import CheckpointConverter as CheckpointConverter
 from fairseq2.models.loader import DelegatingModelLoader as DelegatingModelLoader
 from fairseq2.models.loader import ModelFactory as ModelFactory
@@ -19,7 +20,6 @@ from fairseq2.models.loader import ModelLoader as ModelLoader
 from fairseq2.models.loader import StandardModelLoader as StandardModelLoader
 from fairseq2.models.loader import load_model as load_model
 from fairseq2.models.model import Model as Model
-from fairseq2.models.model import model_factories as model_factories
 
 # isort: split
