+++ src/fairseq2/optim/__init__.py
@@ -9,3 +9,9 @@ from __future__ import annotations
 from fairseq2.optim.adamw import AdamW as AdamW
 from fairseq2.optim.dynamic_loss_scaler import DynamicLossScaler as DynamicLossScaler
 from fairseq2.optim.dynamic_loss_scaler import LossScaleResult as LossScaleResult
+from fairseq2.optim.factory import AdamWConfig as AdamWConfig
+from fairseq2.optim.factory import create_adamw_optimizer as create_adamw_optimizer
+from fairseq2.optim.factory import optimizer_factories as optimizer_factories
+from fairseq2.optim.factory import optimizer_factory as optimizer_factory
+from fairseq2.optim.optimizer import AbstractOptimizer as AbstractOptimizer
+from fairseq2.optim.optimizer import ParameterCollection as ParameterCollection
