+++ src/fairseq2/recipes/mt/translate.py
@@ -20,9 +20,12 @@ from fairseq2.data.text import TextTokenizer, load_text_tokenizer
 from fairseq2.datasets import StaticBatching
 from fairseq2.datasets.text import GenericTextDataset, load_text_dataset
 from fairseq2.gang import Gang
-from fairseq2.generation import Seq2SeqGenerator
-from fairseq2.generation.encoder_decoder import BeamSearchConfig, generator_factories
-from fairseq2.generation.text import SequenceToTextConverter
+from fairseq2.generation import (
+    BeamSearchConfig,
+    Seq2SeqGenerator,
+    SequenceToTextConverter,
+    seq2seq_generator_factories,
+)
 from fairseq2.logging import get_log_writer
 from fairseq2.models import load_model
 from fairseq2.models.encoder_decoder import EncoderDecoderModel
@@ -80,7 +83,7 @@ class TextTranslateConfig:
     """The sequence generator."""
 
     generator_config: Optional[DataClass] = field(
-        default_factory=lambda: BeamSearchConfig()
+        default_factory=lambda: BeamSearchConfig(max_gen_len=(1, 256), echo_prompt=True)
     )
     """The configuration of the sequence generator."""
 
@@ -172,7 +175,7 @@ def load_text_translator(
 
     # Initialize the sequence generator.
     try:
-        generator_factory = generator_factories.get(
+        generator_factory = seq2seq_generator_factories.get(
             config.generator, config.generator_config
         )
